fix: Display objects (instead of string) in console
