+++ apps/hub/src/cli.ts
@@ -1,6 +1,4 @@
 #!/usr/bin/env node
-import { FarcasterNetwork } from '@farcaster/flatbuffers';
-import { HubError, validations } from '@farcaster/utils';
 import { PeerId } from '@libp2p/interface-peer-id';
 import { createEd25519PeerId, createFromProtobuf, exportToProtobuf } from '@libp2p/peer-id-factory';
 import { Command } from 'commander';
@@ -11,7 +9,8 @@ import { exit } from 'process';
 import { APP_VERSION, Hub, HubOptions } from '~/hub';
 import { logger } from '~/utils/logger';
 import { addressInfoFromParts, ipMultiAddrStrFromAddressInfo } from '~/utils/p2p';
-import { startConsole } from './console/console';
+import { DEFAULT_RPC_CONSOLE, parseServerAddress, startConsole } from './console/console';
+import { parseNetwork } from './utils/command';
 
 /** A CLI to accept options from the user and start the Hub */
 
@@ -21,23 +20,13 @@ const DEFAULT_PEER_ID_DIR = './.hub';
 const DEFAULT_PEER_ID_FILENAME = `default_${PEER_ID_FILENAME}`;
 const DEFAULT_PEER_ID_LOCATION = `${DEFAULT_PEER_ID_DIR}/${DEFAULT_PEER_ID_FILENAME}`;
 
-const parseNetwork = (network: string): FarcasterNetwork => {
-  const networkId = Number(network);
-  if (isNaN(networkId)) throw new HubError('bad_request.invalid_param', 'network must be a number');
-  const isValidNetwork = validations.validateNetwork(networkId);
-  if (isValidNetwork.isErr()) {
-    throw isValidNetwork.error;
-  }
-  return isValidNetwork.value;
-};
-
 const app = new Command();
 app.name('hub').description('Farcaster Hub').version(APP_VERSION);
 
 app
   .command('start')
   .description('Start a Hub')
-  .requiredOption('-e, --eth-rpc-url <url>', 'RPC URL of a Goerli Ethereum Node')
+  .option('-e, --eth-rpc-url <url>', 'RPC URL of a Goerli Ethereum Node')
   .option('-c, --config <filepath>', 'Path to a config file with options', DEFAULT_CONFIG_FILE)
   .option('-f, --fir-address <address>', 'The address of the FIR contract')
   .option('-b, --bootstrap <peer-multiaddrs...>', 'A list of peer multiaddrs to bootstrap libp2p')
@@ -189,8 +178,14 @@ app
 app
   .command('console')
   .description('Start a REPL console')
-  .action(async () => {
-    startConsole();
+  .option(
+    '-s, --server <url>',
+    'Farcaster RPC server address:port to connect to (eg. 127.0.0.1:13112)',
+    parseServerAddress,
+    DEFAULT_RPC_CONSOLE
+  )
+  .action(async (cliOptions) => {
+    startConsole(cliOptions.server);
   });
 
 const readPeerId = async (filePath: string) => {
