+++ apps/hub/src/console/castsCommand.ts
@@ -21,84 +21,84 @@ export class CastsCommand implements ConsoleCommandInterface {
       castsByMention <userId> - Get the list of casts by mention`;
   }
 
-  castByFidAndHash = async (fid: string, tsHash: string): Promise<string | MessageModel> => {
+  castByFidAndHash = async (fid: string, tsHash: string): Promise<HubError | MessageModel> => {
     const fidBytes = hexStringToBytes(fid);
     if (fidBytes.isErr()) {
-      return JSON.stringify(fidBytes.error);
+      return fidBytes.error;
     }
 
     const tsHashBytes = hexStringToBytes(tsHash);
     if (tsHashBytes.isErr()) {
-      return JSON.stringify(tsHashBytes.error);
+      return tsHashBytes.error;
     }
 
     const result = await this.rpcClient.getCast(fidBytes._unsafeUnwrap(), tsHashBytes._unsafeUnwrap());
-    return result.match<string | MessageModel>(
+    return result.match<HubError | MessageModel>(
       (cast) => {
         const mm = new MessageModel(cast);
         return mm;
       },
       (err: HubError) => {
-        return `${err}`;
+        return err;
       }
     );
   };
 
-  castsByFid = async (fid: string): Promise<string | MessageModel[]> => {
+  castsByFid = async (fid: string): Promise<HubError | MessageModel[]> => {
     const fidBytes = hexStringToBytes(fid);
     if (fidBytes.isErr()) {
-      return JSON.stringify(fidBytes.error);
+      return fidBytes.error;
     }
 
     const result = await this.rpcClient.getCastsByFid(fidBytes._unsafeUnwrap());
-    return result.match<string | MessageModel[]>(
+    return result.match<HubError | MessageModel[]>(
       (casts) => {
         return casts.map((cast) => new MessageModel(cast));
       },
       (err: HubError) => {
-        return `${err}`;
+        return err;
       }
     );
   };
 
-  castsByParent = async (fid: string, tsHash: string): Promise<string | MessageModel[]> => {
+  castsByParent = async (fid: string, tsHash: string): Promise<HubError | MessageModel[]> => {
     const fidBytes = hexStringToBytes(fid);
     if (fidBytes.isErr()) {
-      return JSON.stringify(fidBytes.error);
+      return fidBytes.error;
     }
 
     const tsHashBytes = hexStringToBytes(tsHash);
     if (tsHashBytes.isErr()) {
-      return JSON.stringify(tsHashBytes.error);
+      return tsHashBytes.error;
     }
 
     const castIdT = new CastIdT(Array.from(tsHashBytes._unsafeUnwrap()), Array.from(tsHashBytes._unsafeUnwrap()));
     const result = await this.rpcClient.getCastsByParent(castIdT);
-    return result.match<string | MessageModel[]>(
+    return result.match<HubError | MessageModel[]>(
       (casts) => {
         return casts.map((cast) => new MessageModel(cast));
       },
       (err: HubError) => {
-        return `${err}`;
+        return err;
       }
     );
   };
 
-  castsByMention = async (userId: string): Promise<string | MessageModel[]> => {
+  castsByMention = async (userId: string): Promise<HubError | MessageModel[]> => {
     const userFidBytes = hexStringToBytes(userId);
     if (userFidBytes.isErr()) {
-      return JSON.stringify(userFidBytes.error);
+      return userFidBytes.error;
     }
 
     const userIdT = new UserIdT(Array.from(userFidBytes._unsafeUnwrap()));
 
     const result = await this.rpcClient.getCastsByMention(userIdT);
-    return result.match<string | MessageModel[]>(
+    return result.match<HubError | MessageModel[]>(
       (casts) => {
         return casts.map((cast) => new MessageModel(cast));
       },
       (err: HubError) => {
-        return `${err}`;
+        return err;
       }
     );
   };
